// projects/x-forms-lib/src/lib/models/variable-format.ts
export enum VariableFormat {
  TEXT = 'TEXT',
  EMAIL = 'EMAIL',
  PHONE = 'PHONE',
  IBAN = 'IBAN',
  BSN = 'BSN',
  NUMBER = 'NUMBER',
  FLOAT = 'FLOAT',
  POSITIVE_NUMBER = 'POSITIVE_NUMBER',
  NEGATIVE_NUMBER = 'NEGATIVE_NUMBER',
  POSITIVE_FLOAT = 'POSITIVE_FLOAT',
  NEGATIVE_FLOAT = 'NEGATIVE_FLOAT',
  AMOUNT = 'AMOUNT',
  DATE = 'DATE',
  DATE_PAST = 'DATE_PAST',
  DATE_FUTURE = 'DATE_FUTURE',
  DATE_PERIOD = 'DATE_PERIOD',
  DATE_BIRTHDAY = 'DATE_BIRTHDAY',
  POSTAL_CODE = 'POSTAL_CODE'
}